requirements.txt
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
curl_cffi==0.6.4
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
idna==3.7
keyboard==0.13.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.1
rsa==4.9
setuptools==70.0.0
SpeechRecognition==3.10.4
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
